Fixed error with markdown in article body.
Used telegram.utils.helpers.escape_markdown to escape any markdown occurring in the body of the text, which causes telegram API to throw an error. Thanks to Jannes Hoke for pointing that out.

For example, http://www.channelnewsasia.com/news/business/2-australian-start-ups-sign-medical-tech-shipping-deals-with-9143160 has a sentence that contains the word A*STAR. * <string> * is used in Markdown to document bold text. Thus, telegram API threw an error as it was unable to find a closing tag for this bold text.